HACK HACK HACK: get firecracker working
The current shim assumes that the VM is up and running at the
time it is launched.

With firecracker we launch the VM late. So the vsock client
connection fails (and fails silently today).

This logic needs to change to wait for the VM to be actually
launched. We cannot depend on a timeout as the shim is launched
at "OCI Create" and the VM is launched at "OCI Start". And the
two commands are fully decoupled from each other.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>